requirements.txt
fastapi>=0.128.0
uvicorn>=0.40.0
openai-whisper>=20250625
pydantic>=2.12.5
requests>=2.31.0